Refactor theme colors and remove unused hooks
Simplified the color variables in the colors.scss file by reducing the variations of darkness and lightness of the theme colors. Also reframed color variable names to a simple numeric system, improving code readability.

Unutilized 'spaces' hook in Sidebar component was also removed, as it was producing unnecessary complexity and was not contributing to any feature of the application.

These modifications aim to