Add ConfigChooserModal and its context
The new ConfigChooserModal component was created to provide a dropdown styled modal for configuration selection. It gets data such as configName, items, various handlers and modal style properties via props. An auxiliary context, ConfigChooserModalContext, was also created for this component, allowing other components to use the modal invoking function directly. It can be useful for avoiding prop drilling. Additionally, Sidebar component was modified to be a parent of ConfigChooserModalProvider which provides this context with the modal invoking function.